backend/requirements.txt
Flask
Flask-Sockets
pymongo
python-dotenv
gevent
gevent-websocket